refactor: 수강신청 (4단계)
